+++ outputs.tf
@@ -1,15 +1,15 @@
 output "ini" {
-  value       = contains(var.formats, "ini") ? local_file.ansible_inventory["ini"].content : ""
+  value       = contains(var.formats, "ini") ? local.inv_content["ini"] : ""
   description = "The Ansible INI format inventory content."
 }
 
 output "yaml" {
-  value       = contains(var.formats, "yaml") ? local_file.ansible_inventory["yaml"].content : ""
+  value       = contains(var.formats, "yaml") ? local.inv_content["yaml"] : ""
   description = "The Ansible YAML format inventory content."
 }
 
 output "json" {
-  value       = contains(var.formats, "json") ? local_file.ansible_inventory["json"].content : ""
+  value       = contains(var.formats, "json") ? local.inv_content["json"] : ""
   description = "The Ansible JSON format inventory content."
 }